pyproject.toml
[project]
name = "zione"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "mflux>=0.5.1",
    "pillow>=10.4.0",
    "pydantic>=2.10.6",
    "uvicorn>=0.34.0",
]

[tool.uv.sources]
mflux = { path = "../mflux" }
